(#1) fix keyboard
